requirements.txt
beautifulsoup4
requests
numpy
pandas
matplotlib
pyyaml
jupyter